Put the test helpers in /test-support in a file with the name of the addon
That way is easier for end users to consume them, because instead of

`import typeInSearch from '<app-name>/tests/helpers/type-in-search';`

they will do

`import { typeInSearch } from '<app-name>/tests/helpers/ember-power-select';`

This is easier to remember and is hard to have names collisions if helpers
have common names like `select` or `choose`.
