+++ tests/integration/components/power-select/general-behaviour-test.js
@@ -1,8 +1,7 @@
 import Ember from 'ember';
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import typeInSearch from '../../../helpers/type-in-search';
-import triggerKeydown from '../../../helpers/trigger-keydown';
+import { typeInSearch, triggerKeydown } from '../../../helpers/ember-power-select';
 import {
   numbers,
   names,
